+++ x/profiles/client/cli/query.go
@@ -77,8 +77,8 @@ func GetCmdQueryProfiles(cdc *codec.Codec) *cobra.Command {
 // GetCmdQueryProfileParams queries all the profiles' module params
 func GetCmdQueryProfileParams(cdc *codec.Codec) *cobra.Command {
 	return &cobra.Command{
-		Use:   "params",
-		Short: "Retrieve all the profile module params",
+		Use:   "parameters",
+		Short: "Retrieve all the profile module parameters",
 		Args:  cobra.ExactArgs(0),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cliCtx := context.NewCLIContext().WithCodec(cdc)
@@ -86,7 +86,7 @@ func GetCmdQueryProfileParams(cdc *codec.Codec) *cobra.Command {
 			route := fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QueryParams)
 			res, _, err := cliCtx.QueryWithData(route, nil)
 			if err != nil {
-				fmt.Printf("Could not find profile params")
+				fmt.Printf("Could not find profile parameters")
 				return nil
 			}