+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ## Bug fixes
 - Fixed a bug that caused valid URIs to be considered invalid (#233)
- 
+- Fixed a bug that didn't allow querying `x/profile` parameters from REST API (#244) 
+
 # Version 0.9.0
 ## Changes
 - Removed the JSON-style `String` methods where it was possible, changed the others to not